feat: add logged out state to fixed header
