+++ .changeset/six-berries-draw.md
@@ -1,0 +1,5 @@
+---
+'@sigle/app': minor
+---
+
+Add a fixed header to the profile page that appears when scrolling down the page.
